adding a new domain to the FileSystemService: AddonScenarioSavegames - these need some special treatment as they are loaded from Addressables

+++ Unity/Assets/GameFramework/Services/FileSystemService/FileSystemServiceImpl.cs
@@ -10,7 +10,7 @@ using Service.Scripting;
 using System.IO.Compression;
 
 namespace Service.FileSystem {
-    partial class FileSystemServiceImpl : FileSystemServiceBase {
+    public partial class FileSystemServiceImpl : FileSystemServiceBase {
 
         public readonly string MISC_PATH = Application.persistentDataPath + "/default";
 
@@ -20,6 +20,7 @@ namespace Service.FileSystem {
         private readonly string streamingAssetsPath = Application.streamingAssetsPath;
         private readonly string persistentDataPath = Application.persistentDataPath;
         private readonly string gamePath =  Application.dataPath.Remove(Application.dataPath.LastIndexOf("/"));
+        private readonly string addonScenarioSavegamesPath = "Assets/Res/Scenarios/ScenarioSavegames/";
 
         private string configPath;
         private string savegamePath;
@@ -103,7 +104,7 @@ namespace Service.FileSystem {
                 case FSDomain.Debugging: path = debuggingPath; break;
                 case FSDomain.Modding: path = moddingPath; break;
                 case FSDomain.SteamingAssets: path = streamingAssetsPath; break;
-
+                case FSDomain.AddonScenarioSavegames: path = addonScenarioSavegamesPath; break;
 
                 default: Debug.LogError("UNKNOWN DOMAIN:" + domain.ToString()+" in GetPath! Using MISC-Path"); break;
             }
@@ -149,6 +150,7 @@ namespace Service.FileSystem {
         }
 
         public override bool WriteStringToFileAtDomain(FSDomain domain, string relativePathToFile, string data,bool append=false) {
+            if (domain == FSDomain.AddonScenarioSavegames) return false;
             relativePathToFile = Utils.CreateValidFilename(relativePathToFile.TrimStart('/'));
             return WriteStringToFile(GetPath(domain, relativePathToFile), data,append);
         }
@@ -189,6 +191,7 @@ namespace Service.FileSystem {
         }
 
         public override bool WriteBytesToFileAtDomain(FSDomain domain, string relativePathToFile, byte[] bytes,bool compress=false) {
+            if (domain == FSDomain.AddonScenarioSavegames) return false;
             relativePathToFile = Utils.CreateValidFilename(relativePathToFile.TrimStart('/'));
             return WriteBytesToFile(GetPath(domain) + "/" + relativePathToFile, bytes,compress);
         }
@@ -213,7 +216,22 @@ namespace Service.FileSystem {
         }
 
         public override byte[] LoadFileAsBytesAtDomain(FSDomain domain, string relativePathToFile, bool compressed = false, int estimatedUncompressedSize = 0) {            
-            return LoadFileAsBytes(GetPath(domain) + "/" + relativePathToFile,compressed,estimatedUncompressedSize);
+
+            if (domain != FSDomain.AddonScenarioSavegames) {
+                return LoadFileAsBytes(GetPath(domain) + "/" + relativePathToFile,compressed,estimatedUncompressedSize);
+            }
+            else {
+
+                string path = GetPath(domain, relativePathToFile);
+                UnityEngine.ResourceManagement.AsyncOperations.AsyncOperationHandle<TextAsset> h = UnityEngine.AddressableAssets.Addressables.LoadAssetAsync<TextAsset>(path);
+                h.WaitForCompletion();
+                byte[] bytes = h.Result.bytes;
+                if (compressed) {
+
+                    return Decompress(bytes);
+                }
+                return bytes;
+            }
         }
 
         public override byte[] LoadFileAsBytes(string pathToFile, bool compressed = false, int estimatedUncompressedSize=0) {
@@ -245,7 +263,17 @@ namespace Service.FileSystem {
         }
 
         public override string LoadFileAsStringAtDomain(FSDomain domain, string relativePathToFile) {
-            return LoadFileAsString(GetPath(domain) + "/" + relativePathToFile);
+
+            if (domain != FSDomain.AddonScenarioSavegames) {
+                return LoadFileAsString(GetPath(domain) + "/" + relativePathToFile);
+            }
+            else {
+
+                string path = GetPath(domain, relativePathToFile);
+                UnityEngine.ResourceManagement.AsyncOperations.AsyncOperationHandle<TextAsset> h = UnityEngine.AddressableAssets.Addressables.LoadAssetAsync<TextAsset>(path);
+                h.WaitForCompletion();
+                return h.Result.text
+            }
         }
 
         public override bool FileExists(string pathToFile) {
@@ -291,7 +319,14 @@ namespace Service.FileSystem {
         }
 
         public override List<string> GetFilesInDomain(FSDomain domain, string innerDomainPath="",string filter = "*.*",bool recursive=false) {
-            return GetFilesInAbsFolder(GetPath(domain,innerDomainPath), filter, recursive);
+            if (domain != FSDomain.AddonScenarioSavegames) {
+
+                return GetFilesInAbsFolder(GetPath(domain,innerDomainPath), filter, recursive);
+            }
+            else {
+                Debug.LogError(typeof(FileSystemServiceImpl).Name + "GetFilesInDomain() is not implemented for domain " + FSDomain.AddonScenarioSavegames.ToString());
+                return null;
+            }
         }
 
         public override void RemoveFile(string filePath) {
@@ -306,6 +341,7 @@ namespace Service.FileSystem {
         }
 
         public override void RemoveFileInDomain(FSDomain domain, string relativePath) {
+            if (domain == FSDomain.AddonScenarioSavegames) return;
             string path = GetPath(domain,relativePath);
             RemoveFile(path);
         }
